[Nightly 2026-01-02] sonamu 체크섬 로직 간소화, tasks PubSub null 체크, 미사용 변수 제거

+++ modules/tasks/src/database/pubsub.ts
@@ -65,9 +65,15 @@ export class PostgresPubSub {
 
   // destroy the listener and close the connection, do not destroy the knex connection
   async destroy() {
-    this._destroyed = true;
-    this._connection.off("close", this._onClosed);
-    await this.knex.client.destroyRawConnection(this._connection);
+    if (this._destroyed) {
+      return;
+    }
+    try {
+      this._connection.off("close", this._onClosed);
+      await this.knex.client.destroyRawConnection(this._connection);
+    } finally {
+      this._destroyed = true;
+    }
   }
 
   // create a new listener and connect to the database
